Revise NRPS GetMembership method
The NRPS specification leaves room for ambiguity in how partial
membership responses may be requested or returned; it does not
specifically allow for offset & limit style paging. This commit makes
a revision that makes the GetMembership use the GetPagedMembership
method as a helper in case partial memberships are returned by the
platform, i.e. a link header with rel="next" is found in the response.