Simplify check for request success
Previously, makeServiceRequest would look for a specific
status code when evaluating whether a request was successful.
The problem: Blackboard Learn and Moodle aren't consistent
in their returning of status codes. For example, putting a
grade with Moodle always returns 200, and putting a grade with
Blackboard Learn might return 201 (when first creating the
grade).

In the future, it may be necessary to reintroduce the idea
of ExpectedStatus and provide it with an array of status
codes. For now, this solution is simpler.